Integrate coverage into categories script
