+++ categories.js
@@ -1,22 +1,40 @@
 "use strict";
-const { rules } = require(".");
+const config = require(".");
 const { Linter } = require("eslint");
+const { readFile } = require("fs/promises");
 const { groupBy } = require("lodash");
 
-const allRules = new Linter().getRules();
-const rulesByCategory = groupBy(
-  Object.keys(rules),
-  (rule) => allRules.get(rule).meta.docs.category
-);
+async function getDocumentedRules() {
+  const markdown = await readFile("README.md", "utf8");
+  const headingMatches = markdown.matchAll(/^#.*\((?<rules>.*)\)$/gmu);
+  return Array.from(headingMatches).flatMap(({ groups: { rules } }) => {
+    const ruleMatches = rules.matchAll(/`(?<rule>[a-z-]+)`/gu);
+    return Array.from(ruleMatches).map(({ groups: { rule } }) => rule);
+  });
+}
 
-console.log(
-  Object.entries(rulesByCategory)
-    .map(([category, categoryRules]) =>
-      [
-        `${category} (${categoryRules.length})`,
-        ...categoryRules.map((rule) => `- ${rule}`),
-      ].join("\n")
-    )
-    .concat(`Total (${Object.keys(rules).length})`)
-    .join("\n\n")
-);
+function getRulesByCategory() {
+  const allRules = new Linter().getRules();
+  return groupBy(
+    Object.keys(config.rules),
+    (rule) => allRules.get(rule).meta.docs.category
+  );
+}
+
+(async () => {
+  const documentedRules = await getDocumentedRules();
+
+  console.log(
+    Object.entries(getRulesByCategory())
+      .map(([category, categoryRules]) =>
+        [
+          `${category} (${categoryRules.length})`,
+          ...categoryRules.map(
+            (rule) => `${documentedRules.includes(rule) ? "✓" : "✖"} ${rule}`
+          ),
+        ].join("\n")
+      )
+      .concat(`Total (${Object.keys(config.rules).length})`)
+      .join("\n\n")
+  );
+})();
